+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceInterface/UserService.java
@@ -2,10 +2,12 @@ package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInter
 
 
 import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDTO;
+import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDetailsDTO;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 
 public interface UserService {
     Page<UserDTO> searchUsers(String searchWord, String role, Long departmentId, Pageable pageable);
+    UserDetailsDTO getUserDetailsById(long id);
 }
