--- src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceInterface/UsersDetailsService.java
@@ -1,7 +1,0 @@
-package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface;
-
-import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDetailsDTO;
-
-public interface UsersDetailsService {
-UserDetailsDTO getUserDetailsById(long id);
-}
